backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
httpx==0.26.0
redis==5.0.1
cassandra-driver>=4.0.0
python-multipart==0.0.9
pyjwt==2.8.0
bcrypt==4.1.2
passlib==1.7.4
python-jose==3.3.0
cryptography==42.0.2
